requirements.txt
torch>=1.9.0
torchvision>=0.10.0
opencv-python
Pillow
numpy
pycocotools
matplotlib
scikit-image
argparse